+++ doc/changelog.rst
@@ -21,6 +21,9 @@ Contributors
 
 Added
 -----
+- Sample tilt about RD can be passed as part of an EBSDDetector. This can be used when
+  projecting parts of master patterns onto a detector.
+  (`#? <https://github.com/pyxem/kikuchipy/pull/?>`_)
 - Unit testing of docstring examples.
   (`#350 <https://github.com/pyxem/kikuchipy/pull/350>`_)
 - Support for Python 3.9.